Extract find query runner partial
